+++ Content.Server/Xenoarchaeology/XenoArtifacts/Triggers/Components/ArtifactGasTriggerComponent.cs
@@ -3,16 +3,25 @@ using Robust.Shared.GameObjects;
 using Robust.Shared.Serialization.Manager.Attributes;
 using Robust.Shared.ViewVariables;
 
-namespace Content.Server.Xenoarchaeology.XenoArtifacts.Triggers;
+namespace Content.Server.Xenoarchaeology.XenoArtifacts.Triggers.Components;
 
+/// <summary>
+///     Activates artifact when it surrounded by certain gas.
+/// </summary>
 [RegisterComponent]
 public class ArtifactGasTriggerComponent : Component
 {
     public override string Name => "ArtifactGasTrigger";
 
+    /// <summary>
+    ///     Should activation gas be picked on startup?
+    /// </summary>
     [DataField("randomGas")]
     public bool RandomGas = true;
 
+    /// <summary>
+    ///     List of possible activation gases to pick on startup.
+    /// </summary>
     [DataField("possibleGas")]
     public Gas[] PossibleGases =
     {
@@ -22,10 +31,16 @@ public class ArtifactGasTriggerComponent : Component
         Gas.CarbonDioxide
     };
 
+    /// <summary>
+    ///     Gas id that will activate artifact.
+    /// </summary>
     [DataField("gas")]
     [ViewVariables(VVAccess.ReadWrite)]
     public Gas? ActivationGas;
 
+    /// <summary>
+    ///     How many moles of gas should be present in room to activate artifact.
+    /// </summary>
     [DataField("moles")]
     [ViewVariables(VVAccess.ReadWrite)]
     public float ActivationMoles = Atmospherics.MolesCellStandard * 0.1f;
